Add module prelude
